API/requirements.txt
affine==2.4.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
blessings==1.7
boto3==1.34.45
botocore==1.34.45
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
earthengine-api==0.1.390
ee==0.2
exceptiongroup==1.2.0
fastapi==0.109.2
geographiclib==2.0
geopy==2.4.1
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
googlemaps==4.10.0
h11==0.14.0
h3==3.7.6
httplib2==0.22.0
idna==3.6
iso3166==2.1.1
jmespath==1.0.1
mapbox==0.18.1
msgpack==1.0.7
numpy==1.26.4
opencage==2.4.0
pillow==10.2.0
polyline==2.0.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
rasterio==1.3.9
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
starlette==0.36.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.2